extract DiscardFileParameters to own file
