+++ Framework/include/QualityControl/TaskRunnerConfig.h
@@ -22,7 +22,7 @@
 
 #include <Framework/DataProcessorSpec.h>
 #include "QualityControl/Activity.h"
-#include "QualityControl/QcInfoLogger.h"
+#include "QualityControl/DiscardFileParameters.h"
 
 namespace o2::quality_control::core
 {
